+++ tests/dtools/fp/state/test_state.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from typing import Final, Never
-from grscheller.fp.state import State
+from dtools.fp.state import State
 
 class Test_simple:
     def test_simple_counter(self) -> None: